+++ enterprise/internal/batches/types/changeset_job.go
@@ -16,13 +16,17 @@ type ChangesetJobCommentPayload struct {
 
 // ChangesetJob describes a one-time action to be taken on a changeset.
 type ChangesetJob struct {
-	ID            int64
+	ID int64
+	// BulkGroup is a random string that can be used to group jobs together in a
+	// single invocation.
 	BulkGroup     string
 	BatchChangeID int64
 	UserID        int32
 	ChangesetID   int64
 	JobType       ChangesetJobType
 	Payload       interface{}
+
+	// workerutil fields
 
 	State          string
 	FailureMessage *string
